refactor: split socket runtime files by integration
