+++ src/runtime/LegacyWDSSocketEntry.js
